refactor: isolate reserved renewal grant activation
Changed:
Moved reserved renewal grant preflight, activation, verification, and ledger balance sync into a dedicated billing service.

Benefit:
Subscription renewal code is easier to maintain while preserving existing credit activation behavior.
